Refactor controllers.
