+++ server/controllers/events/azuredevops_request_validator.go
@@ -1,4 +1,4 @@
-package server
+package events
 
 import (
 	"fmt"
