+++ server/controllers/events/github_request_validator_test.go
@@ -11,7 +11,7 @@
 // limitations under the License.
 // Modified hereafter by contributors to runatlantis/atlantis.
 
-package server_test
+package events_test
 
 import (
 	"bytes"
