+++ server/controllers/events/gitlab_request_parser_validator.go
@@ -11,7 +11,7 @@
 // limitations under the License.
 // Modified hereafter by contributors to runatlantis/atlantis.
 
-package server
+package events_test
 
 import (
 	"encoding/json"
