+++ server/controllers/locks_controller.go
@@ -1,4 +1,4 @@
-package server
+package controllers
 
 import (
 	"fmt"
